Added ability to use cluster processing and cleanup
Added qsub call to 1-fsrecon.sh. The script writes a job script
to working directory and submits it, putting job output in working
directory. This also requires a seperate step 2-fsconvert to convert
and copy the segmenations to the data directory. This in turn requires
bumpping the preprocess script later in the pipeline to 3-process.sh.

Cleanup
* remove fsl5.0- prefix from fsl program names to be more portable
* removed convert.py, not used currently
* assume freesurfer home set already