Refactored 3-preprocess.sh to use job queue.
3-preprocess.sh now loops over subjects and qsubs a separate
script process.sh, passing the toplevel directory and subject
id as variables on the qsub command line.

Assumes that the roi-extract tool is in build/bin subdirectory
and a virtualpython environment is in build/virutalenv. See the
build instructions in the README and the src/build.sh helper
script for more details.